+++ activerecord/test/models/hamster.rb
@@ -1,0 +1,4 @@
+class Hamster < ActiveRecord::Base
+  belongs_to :breeder
+  validates :breeder, presence: true
+end